requirements.txt
django==5.1
django-ninja==1.4.0
django-redis>=6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow>=10.0.0
mozilla-django-oidc>=4.0.0
django-unfold>=0.70.0
django-tailwind==4.4.0
psycopg2-binary>=2.9.9
gunicorn==23.0.0
paramiko==4.0.0
cryptography>=46.0.0
celery>=5.5.0
python-dotenv>=1.2
whitenoise>=6.6
cookiecutter>=2.6
distlib>=0.3.8